Add .vscode and .idea to .gitignore
In this commit, two items have been added to the .gitignore file: .vscode and .idea. This was done to prevent the settings and configurations associated with Visual Studio Code and IntelliJ IDEA from being tracked or versioned by Git. This is normally done to ensure that each individual's settings remain personal. Additionally, the redundant directory src/directus was removed from .gitignore.